+++ liquid/__init__.py
@@ -44,7 +44,7 @@ from .static_analysis import ContextualTemplateAnalysis
 
 from . import future
 
-__version__ = "1.10.3"
+__version__ = "1.11.0"
 
 __all__ = (
     "AwareBoundTemplate",